Implementando a avaliação de credito do cliente

+++ ms-avaliador-credito/src/main/java/com/talissonmelo/modelo/DadosAvaliacao.java
@@ -2,10 +2,8 @@ package com.talissonmelo.modelo;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class DadosAvaliacao {
     private String cpf;
-    private BigDecimal renda;
+    private Long renda;
 }
